% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree_es.R
\name{boost_tree_es}
\alias{boost_tree_es}
\alias{update.boost_tree_es}
\title{General Interface for Boosted Trees with early stopping support}
\usage{
boost_tree_es(
  mode = "unknown",
  mtry = NULL,
  min_n = NULL,
  tree_depth = NULL,
  learn_rate = NULL,
  loss_reduction = NULL,
  sample_size = NULL
)

\method{update}{boost_tree_es}(
  object,
  parameters = NULL,
  mtry = NULL,
  min_n = NULL,
  tree_depth = NULL,
  learn_rate = NULL,
  loss_reduction = NULL,
  sample_size = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible values for this model are "unknown", "regression", or
"classification".}

\item{mtry}{An number for the number (or proportion) of predictors that will
be randomly sampled at each split when creating the tree models (\code{xgboost}
only).}

\item{min_n}{An integer for the minimum number of data points
in a node that are required for the node to be split further.}

\item{tree_depth}{An integer for the maximum deopth of the tree (i.e. number
of splits) (\code{xgboost} only).}

\item{learn_rate}{A number for the rate at which the boosting algorithm adapts
from iteration-to-iteration (\code{xgboost} only).}

\item{loss_reduction}{A number for the reduction in the loss function required
to split further  (\code{xgboost} only).}

\item{sample_size}{An number for the number (or proportion) of data that is
exposed to the fitting routine. For \code{xgboost}, the sampling is done at at
each iteration while \code{C5.0} samples once during traning.}

\item{object}{A boosted tree model specification.}

\item{parameters}{A 1-row tibble or named list with \emph{main}
parameters to update. If the individual arguments are used,
these will supersede the values in \code{parameters}. Also, using
engine arguments in this object will result in an error.}

\item{fresh}{A logical for whether the arguments should be
modified in-place of or replaced wholesale.}

\item{...}{Not used for \code{update()}.}
}
\description{
General Interface for Boosted Trees with early stopping support
}
\examples{
boost_tree_es(mode = "classification", tree_depth = 5)
# Parameters can be represented by a placeholder:
boost_tree_es(mode = "regression", mtry = varying())
model <- boost_tree_es(mtry = 10, min_n = 3)
model
update(model, mtry = 1)
update(model, mtry = 1, fresh = TRUE)

param_values <- tibble::tibble(mtry = 10, tree_depth = 5)

model \%>\% update(param_values)
model \%>\% update(param_values, mtry = 3)

param_values$verbose <- 0
# Fails due to engine argument
# model \%>\% update(param_values)
}
\seealso{
[parsnip::boost_tree()]
}
