% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{fit}
\alias{update}
\alias{translate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{parsnip}{\code{\link[parsnip]{translate}}}

  \item{stats}{\code{\link[stats]{update}}}
}}

