% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree_es.R
\name{xgb_pred_es}
\alias{xgb_pred_es}
\title{Predict by xgboost fitted object with early stopping support}
\usage{
xgb_pred_es(object, newdata, ...)
}
\arguments{
\item{object}{model_fit object}

\item{newdata}{new data.}

\item{...}{Additional arguments for predict().}
}
\description{
Predict by xgboost fitted object with early stopping support
}
\keyword{internal}
