% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree_es.R
\name{xgb_train_es}
\alias{xgb_train_es}
\title{Boosted trees via xgboost with early stopping support}
\usage{
xgb_train_es(
  x,
  y,
  test_x,
  test_y,
  max_depth = 6,
  eta = 0.3,
  colsample_bytree = 1,
  min_child_weight = 1,
  gamma = 0,
  subsample = 1,
  ...
)
}
\arguments{
\item{x}{A data frame or matrix of predictors}

\item{y}{A vector (factor or numeric) or matrix (numeric) of outcome data.}

\item{test_x}{A data frame or matrix of predictors}

\item{test_y}{A vector (factor or numeric) or matrix (numeric) of outcome data.}

\item{max_depth}{An integer for the maximum depth of the tree.}

\item{eta}{A numeric value between zero and one to control the learning rate.}

\item{colsample_bytree}{Subsampling proportion of columns.}

\item{min_child_weight}{A numeric value for the minimum sum of instance
weights needed in a child to continue to split.}

\item{gamma}{An number for the minimum loss reduction required to make a
further partition on a leaf node of the tree}

\item{subsample}{Subsampling proportion of rows.}

\item{...}{Other options to pass to \code{xgb.train}.}
}
\value{
A fitted `xgboost` object.
}
\description{
Boosted trees via xgboost with early stopping support
}
\keyword{internal}
